+++ app/controllers/dorsale/flyboy/application_controller.rb
@@ -1,5 +1,2 @@
 class Dorsale::Flyboy::ApplicationController < ::Dorsale::ApplicationController
-  handles_sortable_columns
-
-  helper ::Dorsale::CustomerVault::ApplicationHelper
 end