+++ app/helpers/dorsale/all_helpers.rb
@@ -11,6 +11,7 @@ module Dorsale::AllHelpers
   include ::Dorsale::RoutesHelper
   include ::Dorsale::TextHelper
   include ::Dorsale::BootstrapHelper
+  include ::Dorsale::SortingHelper
 
   include ::Dorsale::Alexandrie::AttachmentsHelper
   include ::Dorsale::BillingMachine::ApplicationHelper